+++ src/main/java/jokrey/utilities/network/link2peer/node/stream/P2LOrderedInputStream.java
@@ -16,16 +16,16 @@ import java.net.SocketAddress;
 public abstract class P2LOrderedInputStream extends InputStream implements P2LInputStream {
     protected final P2LNodeInternal parent;
     protected final SocketAddress from;
-    protected final int type, conversationId;
-    protected P2LOrderedInputStream(P2LNodeInternal parent, SocketAddress from, int type, int conversationId) {
+    protected final short type, conversationId;
+    protected P2LOrderedInputStream(P2LNodeInternal parent, SocketAddress from, short type, short conversationId) {
         this.parent = parent;
         this.from = from;
         this.type = type;
         this.conversationId = conversationId;
     }
     @Override public SocketAddress getRawFrom() { return from; }
-    @Override public int getType() { return type; }
-    @Override public int getConversationId() { return conversationId; }
+    @Override public short getType() { return type; }
+    @Override public short getConversationId() { return conversationId; }
 
     /**
      * Reads a single byte (represented as an integer between 0-255) from the stream.
